automated

+++ sandbox.py
@@ -6,8 +6,15 @@ import csv
 import cv2
 #import time
 import os
+import subprocess
 import toml
 import numpy as np
+
+import mk_avidemuxpy
+
+def dos_path(path):
+    buf = path.replace('/mnt/','').replace('/','\\\\')
+    return buf[0] + ':' + buf[1:]
 
 def pick_frame(video_path, frame_num):
     cap = cv2.VideoCapture(video_path)
@@ -27,12 +34,17 @@ def frame_check(frame, coord, base_color):
 
 
 config = 'config.toml'
-video = '/mnt/e/編集中Videos/ika/IOHD0573.MP4'
+video = '/mnt/e/編集中Videos/ika/IOHD0417.MP4'
+video_dos = dos_path(video)
+
 jsonfile = f'{video}.json'
 keyframefile = f'{video}_keyframes.txt'
 rangefile = f'{video}_ranges.txt'
 
 conf = toml.load(open(config))
+
+ffprobe = conf['win']['ffprobe']
+avidemux = conf['win']['avidemux']
 
 ssample = conf['image']['start_sample']
 esample = conf['image']['end_sample']
@@ -40,9 +52,11 @@ esample = conf['image']['end_sample']
 if not os.path.isfile(keyframefile):
 
     if not os.path.isfile(jsonfile):
-        print('please search for keyframes')
-        print('ffprobe -show_frames -select_streams v -print_format json video_path > json_path')
-        raise Exception('no json')
+        subprocess.run(
+            f'{ffprobe} -show_frames -select_streams v -print_format json {video_dos} > {jsonfile}',
+            shell=True,
+            text=True
+            )
 
     with open(jsonfile) as f:
         buf = json.load(f)['frames']
@@ -51,7 +65,7 @@ if not os.path.isfile(keyframefile):
     print(f'the num of all frames is {len(buf)}, there are {len(key_frames)} key frames')
 
 
-    # 座標
+    # 座標指定がない場合
     if 'coord' not in conf:
         # サンプル画像がなかったら切り出しをするよう指示
         if not os.path.isfile(ssample):
@@ -113,10 +127,12 @@ if not os.path.isfile(keyframefile):
 
         if frame_check(frame, scoord1, scolor1) < threshold and \
                 frame_check(frame, scoord2, scolor2) < threshold:
-            judges.append('s{0}'.format(fr_info['coded_picture_number']))
+            #judges.append('s{0}'.format(fr_info['coded_picture_number']))
+            judges.append('s{0}'.format(fr_info['pkt_pts_time']))
         elif frame_check(frame, ecoord1, ecolor1) < threshold and \
                 frame_check(frame, ecoord2, ecolor2) < threshold:
-            judges.append('e{0}'.format(fr_info['coded_picture_number']))
+            #judges.append('e{0}'.format(fr_info['coded_picture_number']))
+            judges.append('e{0}'.format(fr_info['pkt_pts_time']))
 
     with open(keyframefile, 'w') as f:
         for j in judges:
@@ -124,26 +140,40 @@ if not os.path.isfile(keyframefile):
 else:
     judges = open(keyframefile, 'r').read().rstrip().splitlines()
 
-mode = ''
-ranges = []
+if not os.path.isfile(rangefile):
+    mode = ''
+    ranges = []
 
-for l in judges:
-    if mode == '' and l[0] == 's':
-        ranges.append([l[1:], ])
-        mode = 's'
-    if mode == 's' and l[0] == 'e':
-        mode = 'e'
-        prev = l[1:]
-    if mode == 'e' and l[0] == 's':
-        ranges[-1].append(prev)
-        ranges.append([l[1:], ])
-        mode = 's'
-ranges[-1].append(prev)
+    for l in judges:
+        if mode == '' and l[0] == 's':
+            ranges.append([l[1:], ])
+            mode = 's'
+        if mode == 's' and l[0] == 'e':
+            mode = 'e'
+            prev = l[1:]
+        if mode == 'e':
+            if l[0] == 's':
+                ranges[-1].append(prev)
+                ranges.append([l[1:], ])
+                mode = 's'
+            else:
+                prev = l[1:]
+    ranges[-1].append(prev)
 
-with open(rangefile, 'w') as f:
-    cobj = csv.writer(f)
-    cobj.writerows(ranges)
-    #for r in ranges:
-    #    f.write(r+'\n')
+    with open(rangefile, 'w') as f:
+        cobj = csv.writer(f)
+        cobj.writerows(ranges)
+        #for r in ranges:
+        #    f.write(r+'\n')
+else:
+    ranges = list(csv.reader(open(rangefile)))
+    print(ranges)
 
 
+script_path = mk_avidemuxpy.mk_avidemuxpy(video, video_dos, ranges)
+script_dos = dos_path(script_path)
+
+proc = subprocess.run(f'"{avidemux}" --run "{script_dos}"',
+    shell = True,
+    text = True
+)
